moved App file to Components dir
